refactor: rename package util -> support
