+++ src/main/java/com/meoguri/linkocean/configuration/security/jwt/JwtAuthenticationFilter.java
@@ -22,7 +22,7 @@ import com.meoguri.linkocean.domain.user.entity.vo.Email;
 import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
 import com.meoguri.linkocean.domain.user.service.UserService;
 import com.meoguri.linkocean.domain.user.service.dto.GetUserResult;
-import com.meoguri.linkocean.util.TokenUtil;
+import com.meoguri.linkocean.support.TokenUtil;
 
 import io.jsonwebtoken.Claims;
 import lombok.extern.slf4j.Slf4j;
