+++ src/main/java/com/meoguri/linkocean/support/querydsl/Join.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.util.querydsl;
+package com.meoguri.linkocean.support.querydsl;
 
 import java.util.List;
