+++ src/main/java/com/meoguri/linkocean/support/querydsl/JoinInfoBuilder.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.util.querydsl;
+package com.meoguri.linkocean.support.querydsl;
 
 import java.util.ArrayList;
 import java.util.Arrays;
